requirements.txt
python-binance==1.0.16
numpy
pandas
streamlit
plotly
prophet